+++ test/octree_find.cpp
@@ -14,7 +14,7 @@ SCENARIO("octree find", "[octree]")
         pcp::octree_parameters_t params;
         params.voxel_grid    = {pcp::point_t{min, min, min}, pcp::point_t{max, max, max}};
         params.node_capacity = node_capacity;
-        params.max_depth     = max_depth;
+        params.max_depth     = static_cast<std::uint8_t>(max_depth);
 
         std::random_device rd;
         std::mt19937 gen(rd());
