+++ test/octree_range_search.cpp
@@ -10,7 +10,7 @@ SCENARIO("range searches on the octree", "[octree]")
     {
         pcp::octree_parameters_t params;
         params.node_capacity = node_capacity;
-        params.max_depth     = max_depth;
+        params.max_depth     = static_cast<std::uint8_t>(max_depth);
         params.voxel_grid    = pcp::axis_aligned_bounding_box_t{
             pcp::point_t{-1.f, -1.f, -1.f},
             pcp::point_t{1.f, 1.f, 1.f}};
